requirements.txt
Flask~=2.3.2
Flask-Cors==3.0.10
keras~=2.13.1rc0
tensorflow~=2.13.0rc0
numpy~=1.23.5
Pillow~=9.5.0
gevent~=22.10.2
